Restructure moleculess
