+++ src/components/molecule/HomeHeader/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+const styles = () =>
+  StyleSheet.create({
+    midContent: { flex: 1, justifyContent: "center" },
+    container: {
+      flexDirection: "row",
+    },
+  });
+
+export default styles;
