+++ src/components/organism/CheckBoxes/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+import { spacing as sp } from "../../../constants";
+
+const styles = () =>
+  StyleSheet.create({
+    container: { marginVertical: sp.sm },
+  });
+
+export default styles;
