+++ src/components/organism/DrawerContent/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+import { spacing as sp } from "../../../constants";
+
+const styles = () =>
+  StyleSheet.create({
+    buttonContainer: { paddingVertical: sp.s },
+  });
+
+export default styles;
